refactor: improve post view with better auth handling, error states, and UI
